client/js/connection.js: move addEvents2Forward outside
