+++ client/js/presenter.js
@@ -79,6 +79,24 @@ this.readSessionInfo = function(){
 }
 
 this.connect = function(){
+  var events2Forward = [
+    "asq:sessionFlow",
+    "asq:folo-connected",
+    "asq:ctrl-connected",
+    "asq:ghost-connected",
+    "asq:connected-clients",
+    "asq:answered-all",
+    "asq:user-session-stats",
+    "asq:rankings",
+    "asq:goto",
+    "asq:submitted",
+    "asq:assessment",
+    "asq:assess",
+    "asq:stat",
+    "asq:question_type",
+    "asq:session-terminated",
+  ];
+  connection.addEvents2Forward(events2Forward);
   connection.connect(this.protocol, this.host, this.port, this.sessionId, this.namespace, this.token, eventBus);
 }
